用户中心

+++ user/infrastructure/data/UserDO.go
@@ -4,7 +4,7 @@ import "time"
 
 type UserDO struct {
     Id        int
-    UserId    int       `xorm:"notnull unique(uk_user_user_id)"`
+    UserId    int64     `xorm:"notnull unique(uk_user_user_id)"`
     Username  string    `xorm:"varchar(20) notnull"`
     Nickname  string    `xorm:"varchar(20) notnull"`
     Password  string    `xorm:"varchar(60) notnull"`
